Complete Cart Item Controller

+++ Vastra.API/Vastra.API/Controllers/OrdersController.cs
@@ -6,6 +6,7 @@ using Microsoft.EntityFrameworkCore;
 using System.Xml.Linq;
 using Vastra.API.DBContexts;
 using Vastra.API.Entities;
+using Vastra.API.Enums;
 using Vastra.API.Models;
 using Vastra.API.Models.ForCreationAndUpdate;
 using Vastra.API.Services;
@@ -102,6 +103,7 @@ namespace Vastra.API.Controllers
 
             finalorder.DateAdded = DateTime.Now;
             finalorder.DateModified = DateTime.Now;
+            finalorder.PaymentStatus = PaymentStatus.Pending.ToString();
             await _vastraRepository.AddOrderForUserAsync(userId, finalorder);
             await _vastraRepository.SaveChangesAsync();
 
@@ -144,7 +146,7 @@ namespace Vastra.API.Controllers
         }
 */
 
-        [HttpDelete]
+        [HttpDelete("{orderId}")]
         [Authorize]
         public async Task<ActionResult> DeleteOrder(int roleId, int userId, int orderId)
         {
@@ -164,6 +166,11 @@ namespace Vastra.API.Controllers
             if(orderToDelete == null)
             {
                 return NotFound();
+            }
+            //if order has been placed and payment done, order can't be deleted
+            if (orderToDelete.PaymentStatus.Equals(PaymentStatus.Success.ToString()))
+            {
+                return BadRequest();
             }
             _vastraRepository.DeleteOrder(orderToDelete);
             await _vastraRepository.SaveChangesAsync();
